quizapplication/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
black==23.3.0
cffi==1.15.1
click==8.1.3
Django==4.2.1
django-crispy-forms==1.14.0
django-environ==0.10.0
django-mathfilters==1.0.0
django-utils-six==2.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
psycopg2-binary==2.9.6
pycparser==2.21
sqlparse==0.4.4
python-dotenv==1.0.0